+++ brainiak/utils/fmrisim.py
@@ -509,8 +509,7 @@ def generate_stimfunction(onsets,
 
     # Check files
     if np.max(onsets) > total_time:
-        logger.info('Onsets outside of range of total time. Aborting')
-        exit()
+        raise ValueError('Onsets outside of range of total time.')
 
     # Generate the time course as empty, each element is a millisecond by
     # default